PID tuning. Dropping fixed (mostly). Calibration stuff in place, but currently unused.
Calibration does not actually solve our problems, as the current
implementation only addresses one type of error (sag from gravity,
which is always approximately the same for a given arm position).
The error from motion in a certain direction (over/undershoot) is
NOT modeled, and thus cannot be fixed.

The error we've actually fixed is in the drop command, which did not
take into account the dimensions of the object it was holding when
placing it. The gripper modeled the object has having its centroid
aligned with the theoretical center of the gripper used for
planning, but this is often not the case, hence failures in dropping.