Counting test can handle noisy detections.
There's a hard threshold that should probably be parameterizable
which determines a cutoff for detection "quality". A penalty is applied
based on the number of samples observed so far such that more than one
observation is necessary to ever confirm having "counted" something, and
it is possible to even have the counter realize that an object should
NOT have been counted if later bad samples are observed.