First working DFS.
Some hacks in here to make it terminate when in finds the goal. Greedily
chooses first successful reaching of goal instead of trying to any sort
of 'optimality.' No real ways to deal with noise, yet, but we are
starting to incorporate sampling. Needs to be iteratively deepening, I
think.