First pass at incorporating better early pruning metric.
Use larger number of tag samples to estimate failure rate early, apply
this when actually planning. Our current problems are:

1) We still go down the wrong paths too long and aggressively. If we
choose badly for our first direction, we pay a search price since we are
unable to prune.
2) In general, I'm unsure if we're at our max ability to prune, and
additionally, it feels like we should be able to do better. We also need
to revist our scoring metric anyway since it clearly has a bit of
jankiness left which speeds up search at the cost of formality.