Fixed tag localization and scanmatching.
Produced a new map of CSE which is largely useable. Certainly enough for
some short term testing. Issues detecting tags in log used to generate
the map render the map impractical to use in the final version, but we
should be able to address this fairly easily.

Scanmatching error related to coordinates of the points being matched.
Points were incorrectly pre-transformed instead of being placed in robot
local frame. This resulted in the points being transformed an extra time
inside the scanmatcher. Parameters may still be tuning, but resulting
maps seem adequate.