Fixed control law bug with exploding LCM subscriptions.
When controls laws were created, they also subscribed to relevant
channels. Over time, this would continue until eventually, LCM would
silently blow up and we would stop receiving traffic (at least in sim).
This is obviously not good. Now, when messages are set to run, they
subscribe. When they are set to NOT run (which happens when they are
terminated), they unsubscribe. Current, there is not equivalent change
for ConditionTests, but a similar bug exists and needs to be fixed.